"Default spaces" setting has been moved to "Space Settings".
