+++ protected/humhub/modules/admin/models/forms/SpaceSettingsForm.php
@@ -2,6 +2,7 @@
 
 namespace humhub\modules\admin\models\forms;
 
+use humhub\modules\space\models\Space;
 use Yii;
 
 /**
@@ -14,6 +15,8 @@ class SpaceSettingsForm extends \yii\base\Model
     public $defaultVisibility;
     public $defaultJoinPolicy;
     public $defaultContentVisibility;
+    public $defaultSpaceGuid = [];
+    public $defaultSpaces;
 
     private $settings;
 
@@ -25,6 +28,7 @@ class SpaceSettingsForm extends \yii\base\Model
         $this->defaultJoinPolicy = $this->getSettings()->get('defaultJoinPolicy');
         $this->defaultVisibility = $this->getSettings()->get('defaultVisibility');
         $this->defaultContentVisibility = $this->getSettings()->get('defaultContentVisibility');
+        $this->defaultSpaces = \humhub\modules\space\models\Space::findAll(['auto_add_new_members' => 1]);
     }
 
     private function getSettings()
@@ -42,6 +46,7 @@ class SpaceSettingsForm extends \yii\base\Model
     {
         return [
             [['defaultVisibility', 'defaultJoinPolicy', 'defaultContentVisibility'], 'integer'],
+            ['defaultSpaceGuid', 'checkSpaceGuid'],
         ];
     }
 
@@ -53,10 +58,30 @@ class SpaceSettingsForm extends \yii\base\Model
     public function attributeLabels()
     {
         return [
+            'defaultSpaceGuid' => Yii::t('AdminModule.forms_SpaceSettingsForm', 'Default space'),
             'defaultVisibility' => Yii::t('AdminModule.forms_SpaceSettingsForm', 'Default Visibility'),
             'defaultJoinPolicy' => Yii::t('AdminModule.forms_SpaceSettingsForm', 'Default Join Policy'),
             'defaultContentVisibility' => Yii::t('AdminModule.forms_SpaceSettingsForm', 'Default Content Visiblity'),
         ];
+    }
+
+    /**
+     * This validator function checks the defaultSpaceGuid.
+     * @param type $attribute
+     * @param type $params
+     */
+    public function checkSpaceGuid($attribute, $params)
+    {
+        if (!empty($this->defaultSpaceGuid)) {
+            foreach ($this->defaultSpaceGuid as $spaceGuid) {
+                if ($spaceGuid != "") {
+                    $space = \humhub\modules\space\models\Space::findOne(['guid' => $spaceGuid]);
+                    if ($space == null) {
+                        $this->addError($attribute, Yii::t('AdminModule.forms_SpaceSettingsForm', "Invalid space"));
+                    }
+                }
+            }
+        }
     }
 
     public function save()
@@ -64,7 +89,31 @@ class SpaceSettingsForm extends \yii\base\Model
         $this->getSettings()->set('defaultJoinPolicy', $this->defaultJoinPolicy);
         $this->getSettings()->set('defaultVisibility', $this->defaultVisibility);
         $this->getSettings()->set('defaultContentVisibility', $this->defaultContentVisibility);
+        $this->updateDefaultSpaces();
         return true;
+    }
+
+    private function updateDefaultSpaces()
+    {
+        // Remove Old Default Spaces
+        if (empty($this->defaultSpaceGuid)) {
+            Space::updateAll(['auto_add_new_members' => 0], ['auto_add_new_members' => 1]);
+        } else {
+            foreach (Space::findAll(['auto_add_new_members' => 1]) as $space) {
+                if (!in_array($space->guid, $this->defaultSpaceGuid)) {
+                    $space->auto_add_new_members = 0;
+                    $space->save();
+                }
+            }
+            // Add new Default Spaces
+            foreach ($this->defaultSpaceGuid as $spaceGuid) {
+                $space = Space::findOne(['guid' => $spaceGuid]);
+                if ($space != null && $space->auto_add_new_members != 1) {
+                    $space->auto_add_new_members = 1;
+                    $space->save();
+                }
+            }
+        }
     }
 
 }
